+++ tests/test_cli.py
@@ -7,7 +7,7 @@ import os.path
 import pytest
 from click.testing import CliRunner
 
-from buchfink.cli import fetch_, init, report_
+from buchfink.cli import fetch_, init, report_, quote
 
 logger = logging.getLogger(__name__)
 
@@ -44,3 +44,17 @@ def test_second_init_should_fail():
         result = runner.invoke(init)
         logger.debug('output of %s: %s', 'init', result.output)
         assert result.exit_code == 1
+
+
+def test_init_and_subsequent_quote():
+    runner = CliRunner()
+    with runner.isolated_filesystem() as d:
+        assert os.path.exists(d)
+        result = runner.invoke(init)
+        logger.debug('output of %s: %s', 'init', result.output)
+        assert result.exception is None
+        assert result.exit_code == 0
+        result = runner.invoke(quote, ['ETH', '-b', 'USD'])
+        logger.debug('output of %s: %s', 'init', result.output)
+        assert result.exit_code == 0
+        assert result.exception is None